+++ slcpy.py
@@ -1,1 +1,10 @@
 """Main module."""
+from slcpy import *
+
+
+def main():
+    return
+
+
+if __name__ == '__main__':
+    main()